Added in a couple more test level components, updated the level controller and pool manager scripts along with re added the randomness to the spawners.
Two new level components, large hole start and large hole finish. These were created as two separate components as in the future a large hole may have different components that it can go to. Level controller class now controls the speed and velocity of all the level components, updated the level component script to allow this. Uncommented the randomness in the spawner and updated it to take into consideration that some components only have certain components following them. Added in code to the poolmanager to allow scripts to grab all the names of the level components.